backend/app/agents/requirements.txt
# SNEL DeFi Agent - Coral Protocol Requirements
# ENHANCEMENT FIRST: Extends existing SNEL with Coral capabilities

# Core dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
toml>=0.10.0

# Optional MCP Protocol (install when available)
# mcp>=0.1.0
# langchain-mcp-adapters>=0.1.0

# Existing SNEL dependencies (install from main requirements.txt)
# web3>=6.0.0
# requests>=2.28.0
# fastapi>=0.100.0
# uvicorn>=0.23.0
